On second thought keep the check so that PUT behaves as expected (i.e. preserving IDs) even though I don't think we're using it at the moment

+++ src/main/java/edu/oakland/service/ProfilePictureService.java
@@ -52,7 +52,9 @@ public class ProfilePictureService {
             throw new IllegalArgumentException();
         }
         User user = userOption.get();
-        profilePictureRepository.deleteByUser(user);
+        if (profilePicture.getId() == null) {
+            profilePictureRepository.deleteByUser(user);
+        }
         profilePicture = profilePictureRepository.save(profilePicture);
         user.setImageUrl("/content/user-profile-picture/" + profilePicture.getId());
         user = userRepository.save(user);
